refactor(swarm-storer): move chunk store behind the Database trait
Replace the redb-specific RedbChunkStore with DbChunkStore, generic over
the vertex-storage Database trait, so the storer shares the node's single
database handle and inherits the in-memory or on-disk choice made at open
time. The direct redb dependency and its error conversions are dropped.

The constructor ensures the chunk table exists, so read paths work on a
fresh database without a separate initialization step. for_each iterates
keys only with early exit, and the presence probes in contains and put
are documented against the follow-up trait primitives tracked in #214.
